gateway/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
grpcio==1.66.1
grpcio-tools==1.66.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
protobuf==5.28.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-mock==3.14.0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
